+++ examples/remoteCSE.py
@@ -24,7 +24,7 @@ if __name__ == '__main__':
 	session = Session(conf.CSEURL, conf.ORIGINATOR)
 
 	# Get the <CSEBase> resource
-	cse = CSEBase(session, conf.CSEID)
+	cse = session.getCSEBase()
 
 	# Get a list of all <remoteCSE> resource
 	remoteCSEs = cse.remoteCSEs()